+++ forge/src/main/java/traben/entity_model_features/forge/mixin/MixinLivingEntityRenderer.java
@@ -71,6 +71,11 @@ public abstract class MixinLivingEntityRenderer<T extends LivingEntity, M extend
         if (EMFConfig.getConfig().vanillaModelRenderMode != EMFConfig.VanillaModelRenderMode.Off){
             EMFManager.getInstance().tryRenderVanillaRoot(emf$ModelId,matrixStack,vertexConsumerProvider.getBuffer(RenderLayer.getEntityTranslucent(getTexture(livingEntity))),i, OverlayTexture.DEFAULT_UV);
         }
+        //simple attempt at a physics mod workaround
+        if(livingEntity.isDead() && EMFConfig.getConfig().attemptPhysicsModPatch_1 && EMFManager.getInstance().physicsModInstalled){
+            EMFManager.getInstance().tryRenderVanillaRootNormally(emf$ModelId,matrixStack,vertexConsumerProvider.getBuffer(RenderLayer.getEntityTranslucent(getTexture(livingEntity))),i, OverlayTexture.DEFAULT_UV);
+            //the regular render will get cancelled anyway nothing further to do
+        }
     }
 
     @Inject(method = "render(Lnet/minecraft/entity/LivingEntity;FFLnet/minecraft/client/util/math/MatrixStack;Lnet/minecraft/client/render/VertexConsumerProvider;I)V",